unify Victor serial processes
